refactor(core): keep hook provisioning profile in experience
